Working upload, new button not functional
Working upload but new button to upload photo not working
