+++ app/src/main/res/layout/fragment_image_with_data.xml
@@ -7,31 +7,11 @@
     tools:context=".ImageWithDataFragment">
 
     <TextView
-        android:id="@+id/timerTotal"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Total Time: 00:00"
-        app:layout_constraintTop_toBottomOf="@+id/lblPredictedAmount"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginTop="16dp" />
-
-    <TextView
-        android:id="@+id/timerInterval"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Interval Time: 00:00"
-        app:layout_constraintTop_toBottomOf="@+id/timerTotal"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:layout_marginTop="8dp" />
-
-    <TextView
         android:id="@+id/lblPredictedAmount"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="64dp"
-        android:text="Data simulator"
+        android:text="@string/get_amount_in_image"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.498"
         app:layout_constraintStart_toStartOf="parent"
@@ -53,7 +33,7 @@
         android:layout_width="90dp"
         android:layout_height="wrap_content"
         android:layout_marginBottom="16dp"
-        android:text="Sim"
+        android:text="@string/upload_new"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
@@ -81,65 +61,4 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.136" />
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="48dp"
-        android:text="Set simulation timer:"
-        app:layout_constraintBottom_toTopOf="@+id/inputSimulationTimer"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="40dp"
-        android:text="Set simulation interval:"
-        app:layout_constraintBottom_toTopOf="@+id/inputSimulationInterval"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <EditText
-        android:id="@+id/inputSimulationTimer"
-        android:layout_width="130dp"
-        android:layout_height="42dp"
-        android:ems="10"
-        android:hint="minutes"
-        android:inputType="text"
-        android:textAlignment="center"
-        app:layout_constraintBottom_toTopOf="@+id/guideline"
-        app:layout_constraintEnd_toStartOf="@+id/inputSimulationInterval"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <EditText
-        android:id="@+id/inputSimulationInterval"
-        android:layout_width="130dp"
-        android:layout_height="wrap_content"
-        android:ems="10"
-        android:hint="minutes"
-        android:inputType="text"
-        android:textAlignment="center"
-        app:layout_constraintBottom_toTopOf="@+id/guideline"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toEndOf="@+id/inputSimulationTimer" />
-
-    <androidx.constraintlayout.widget.Guideline
-        android:id="@+id/guideline"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        app:layout_constraintGuide_percent="0.89" />
-
-    <TextView
-        android:id="@+id/simRestaurantName"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/placeholder"
-        android:textSize="28sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
 </androidx.constraintlayout.widget.ConstraintLayout>
